moved models with server. added zipcode column to user table
